% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alpha_function}
\alias{alpha_function}
\title{Forward Probability function}
\usage{
alpha_function(m, N, sigma, Gamma, p, set)
}
\arguments{
\item{m}{number of Likelihoods}

\item{N}{length of the supplied dataset}

\item{sigma}{sigma vector}

\item{Gamma}{Gamma matrix}

\item{p}{vector of Likelihood probabilities of dataset}

\item{set}{index vector to align the p vector}
}
\value{
Returns the alpha as a matrix with the dimension m as columns 
and the datapoints t as rows
}
\description{
This function calculates the forward probability of the HMM states. In the general 
literatur also revered to as alpha matrix
}
