% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{anyNameButFunctionNameIsUnique}
\alias{anyNameButFunctionNameIsUnique}
\alias{alsoToBar}
\alias{alsoNotToBar}
\alias{alsoTheQuestion}
\title{Group of functions page title}
\usage{
alsoToBar(x)

alsoNotToBar(x, y)

alsoTheQuestion(z)
}
\arguments{
\item{x}{a param for toBar and notToBar}

\item{y}{a param just for notToBar}

\item{z}{a param just for theQuestion.

Could put here, but easiest if all params are
described in the same place, with page documentation, so none get repeated.}
}
\value{
Hard to have one return section for all functions,
might want to have a manual list here.
}
\description{
Group of functions Description section
}
\details{
Group of functions Details paragraph.
}
\section{After Arguments and Value sections}{

Despite its location, this actually comes after the Arguments and Value sections.
Also, don't need to use null, could annotate first function, and then
using function name as the groupBy name is more intuitive.
}

\section{Another section}{

Probably better if all sections come first, uless have one section per function. Makes it easier to
see the information flow.
}

