% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_probabilities.R
\name{backward_probabilities}
\alias{backward_probabilities}
\title{backward_probabilities}
\usage{
backward_probabilities(x, Gamma, p)
}
\arguments{
\item{x}{Vector of observed emissions}

\item{Gamma}{Matrix (m x m) of transition probabilities of the underlying Markov chain}

\item{p}{A function where p(i, x)=P(X_t=x | Y_t=i)}
}
\value{
A matrix (m x n) of backward probabilities.
}
\description{
Calculate matrix of backward probabilities for a given data set and parameters in a HMM. Essentially just a wrapper for the true function written in C++.
}
\details{
Let (X, Y)_i, i=1,...,n constitute a HMM, that is, Y_1,...,Y_n is a Markov chain with states 1,...,m and the X_i's are dependent only through the Y_i's with common conditional distributions X_i | Y_i=j = P_j for all i=1,...,n. This function then generates a matrix where the i,j'th entry is P(X_j+1=x_j+1, ..., X_n=x_n | Y_j=i), where x_1,...,x_n are the observed values (emissions) of X.
}
\examples{
TODO
}
