% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta_function}
\alias{beta_function}
\title{Backward Probability function}
\usage{
beta_function(m, N, Gamma, p, set)
}
\arguments{
\item{m}{number of Likelihoods}

\item{N}{length of the supplied dataset}

\item{Gamma}{Gamma matrix}

\item{p}{vector of Likelihood probabilities of dataset}

\item{set}{index vector to align the p vector}
}
\value{
Returns the beta as a matrix with the dimensions m rows
and the datapoints t as collumns
}
\description{
This function calculates the backward probability of the HMM states. In the general 
literatur also revered to as beta matrix
}
