% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{calcAccuracy}
\alias{calcAccuracy}
\title{Calculate the accuracy of a model}
\usage{
calcAccuracy(testresult)
}
\arguments{
\item{testresult}{A matrix containing actual and predicted values as returned by the function test_model().}
}
\value{
A percentage value between 0 and 1. 0.5 means that 50% of the predictions were correct.
}
\description{
The accuracy is the fraction of predictions that were true. It is calculated by dividing the number of correct predictions through
the total number of predictions.
}
