% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{create_time_sequence}
\alias{create_time_sequence}
\title{Creates a STS time sequence out of jam genre data}
\usage{
create_time_sequence(clearedData)
}
\arguments{
\item{clearedData}{A matrix of jams with genres as returned by the function add_genres_to_jams()}
}
\value{
sequences in STS format
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kept.
}
\examples{
# Example for the case that all datasets are located in S:/
jams <- load_jams("S:/")
jams_with_genre <- add_genres_to_jams(jams, "S:/")
jam_sequence <- create_time_sequence(jams_with_genre)
}
