% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_functions.R
\name{decode}
\alias{decode}
\title{Local and global decoding}
\usage{
decode(x, m, L1, L2, L3 = NULL, L4 = NULL, L5 = NULL, gamma, delta,
  theta, multi = FALSE)
}
\arguments{
\item{x}{a sample of a Hidden Markov Model}

\item{m}{the number of states}

\item{L1}{likelihood of the first hidden state}

\item{L2}{likelihood of the second hidden state}

\item{L3}{optional. likelihood of the third hidden state}

\item{L4}{optional. likelihood of the 4th hidden state}

\item{L5}{optional. likelihood of the 5th hidden state}

\item{gamma}{estimated Gamma matrix from previous estimation}

\item{delta}{estimated Delta vector from previous estimation}

\item{theta}{estimated Theta values from previous estimation}

\item{multi}{parameter. TRUE if multifactor HMM() is used}
}
\value{
Returns the path of the local and global decoding as well as a
statement, whether the two paths differ.
}
\description{
Calculation of local and global decoding given the dataset and
the coresponding parameters of the HMM
}
