% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_binom.R
\name{em.binom}
\alias{em.binom}
\title{em.binom}
\usage{
em.binom(obs, gamma, delta, n, p, ...)
}
\arguments{
\item{p}{}
}
\value{

}
\description{
A wrapper for the em function used for binomial data. Should generally not be called by the user.
}
