% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_exponential.R
\name{em.exponential}
\alias{em.exponential}
\title{em.exponential}
\usage{
em.exponential(obs, gamma, delta, lambda, ...)
}
\arguments{
\item{lambda}{}
}
\value{

}
\description{
A wrapper for the em function used for exponential data. Should generally not be called by the user.
}
