% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_normal.R
\name{em.normal}
\alias{em.normal}
\title{em.normal}
\usage{
em.normal(obs, gamma, delta, mu, sigma, ...)
}
\arguments{
\item{sigma}{}
}
\value{

}
\description{
A wrapper for the em function used for normal data. Should generally not be called by the user.
}
