% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_poisson.R
\name{em.poisson}
\alias{em.poisson}
\title{em.poisson}
\usage{
em.poisson(obs, gamma, delta, lambda, ...)
}
\arguments{
\item{lambda}{}
}
\value{

}
\description{
A wrapper for the em function used for Poisson data. Should generally not be called by the user.
}
