% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EvaluationBackward}
\alias{EvaluationBackward}
\title{Evaluation problem - Backward algorithm}
\usage{
EvaluationBackward(A, B, P, O)
}
\arguments{
\item{A}{initial transition probability}

\item{B}{initial observation probability}

\item{P}{initial probability distribution}

\item{O}{observation sequence}
}
\value{
Observation sequence probability
}
\description{
Evaluation problem - Backward algorithm
}
