% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_methods.R
\name{fitted.hmm}
\alias{fitted.hmm}
\title{Fitted values for Hidden Markov Models}
\usage{
\method{fitted}{hmm}(x, method = "local", ...)
}
\arguments{
\item{x}{Object of type \code{hmm}.}

\item{method}{Method for state prediction. One of either 'local' (default, resulting in local decoding), 'viterbi' or 'global' (both resulting in global decoding using Viterbi's algorithm).}

\item{...}{Other arguments.}
}
\value{
List of predicted states and emission values.
}
\description{
Outputs predicted values of Hidden Markov Model in the form of a list where the first element is a vector of state predictions and second element is a vector of emission predictions (expected value corresponding predicted state).
}
\details{
If the distribution family is one of the included standard families (Poisson, normal, binomial or exponential), the expected value will be the plug-in estimate using the fitted parameters of each state.
Otherwise, the empirical mean will be used as the expected value.
}
\examples{
# Continuation of Earthquake example
\dontshow{example(hmm)}
fitted.values(hmm.EQ)
}
