% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forward_ll_cpp}
\alias{forward_ll_cpp}
\title{forward_ll_cpp}
\usage{
forward_ll_cpp(Gamma, p, delta)
}
\arguments{
\item{Gamma}{A matrix of transition log-probabilities for the underlying Markov chain}

\item{p}{A matrix of observation log-probabilites, i.e. where the i,j'th entry is the log-likelihood of observing x_j given the hidden state is i}

\item{delta}{A vector of starting log-probabilites, i.e. the log of the distribution of the first hidden state}
}
\description{
Calculate matrix of forward log-probabilities for a given data set and parameters in a HMM
}
