% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_logprobabilities.R
\name{forward_logprobabilities}
\alias{forward_logprobabilities}
\title{forward_logprobabilities}
\usage{
forward_logprobabilities(x, Gamma, p, delta)
}
\arguments{
\item{x}{Vector of observed emissions}

\item{Gamma}{Matrix (m x m) of transition probabilities of the underlying Markov chain}

\item{p}{A function where p(i, x)=P(X_t=x | Y_t=i)}

\item{delta}{Vector of starting probabilities, i.e. the distribution of Y_1}
}
\value{
A matrix (m x n) of forward log-probabilities.
}
\description{
Calculate matrix of forward log-probabilities for a given data set and parameters in a HMM. Essentially just a wrapper for the true function written in C++. Useful for avoiding underflow (by using the log-exp-sum-trick) when the HMM is large.
}
\details{
Let (X, Y)_i, i=1,...,n constitute a HMM, that is, Y_1,...,Y_n is a Markov chain with states 1,...,m and the X_i's are dependent only through the Y_i's with common conditional distributions X_i | Y_i=j = P_j for all i=1,...,n. This function then generates a matrix where the i,j'th entry is log P(X_1=x_1, ..., X_j=x_j, Y_j=i), where x_1,...,x_n are the observed values (emissions) of X.
}
\examples{
TODO
}
