% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forward_prob_cpp}
\alias{forward_prob_cpp}
\title{forward_prob_cpp}
\usage{
forward_prob_cpp(Gamma, p, delta)
}
\arguments{
\item{Gamma}{A matrix of transition probabilities for the underlying Markov chain}

\item{p}{A matrix of observation probabilites, i.e. where the i,j'th entry is the likelihood of observing x_j given the hidden state is i}

\item{delta}{A vector of starting probabilites, i.e. the distribution of the first hidden state}
}
\description{
Calculate matrix of forward probabilities for a given data set and parameters in a HMM
}
