\name{hmm-package}
\alias{hmm-package}
\alias{hmm}
\docType{package}
\title{
\packageTitle{hmm}
}
\description{
\packageDescription{hmm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hmm}
\packageIndices{hmm}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{hmm}

Maintainer: \packageMaintainer{hmm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
