% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMModel.R
\name{HMMModel}
\alias{HMMModel}
\title{Create a Hidden Markov Model}
\description{
Create a Hidden Markov Model

Create a Hidden Markov Model
}
\details{
A HMM has an identifier, states and observable signals
}
\examples{

## ------------------------------------------------
## Method `HMMModel$setStateTransitions`
## ------------------------------------------------

M <- HMMModel$new("Ex", c("state1", "state2"), c("signalA", "signalB", "signalC"))
M$setStateTransitions(matrix(c(0.4, 0.6, 0.6, 0.4), byrow = TRUE, ncol = 2))
M$state.transitions

## ------------------------------------------------
## Method `HMMModel$setSignalProbability`
## ------------------------------------------------

M <- HMMModel$new("Ex", c("state1", "state2"), c("signalA", "signalB", "signalC"))
M$setSignalProbability(matrix(c(0.4, 0.4, 0.2, 0.1, 0.1, 0.8), byrow = TRUE, ncol = 3))
M$signal.probabilities
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{identifier}}{Name given to this model}

\item{\code{states}}{States this HMM can be in}

\item{\code{signal.set}}{Observable signals that this model can produce}

\item{\code{state.transitions}}{Probabilities for transitions between states}

\item{\code{signal.probabilities}}{Probabilities to produce signals in certain states}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HMMModel$new()}}
\item \href{#method-setStateTransitions}{\code{HMMModel$setStateTransitions()}}
\item \href{#method-setSignalProbability}{\code{HMMModel$setSignalProbability()}}
\item \href{#method-print}{\code{HMMModel$print()}}
\item \href{#method-clone}{\code{HMMModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new HMM object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HMMModel$new(identifier, states, signal.set)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{identifier}}{Identifier.}

\item{\code{states}}{States.}

\item{\code{signal.set}}{Signal.set.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `HMM` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setStateTransitions"></a>}}
\if{latex}{\out{\hypertarget{method-setStateTransitions}{}}}
\subsection{Method \code{setStateTransitions()}}{
Set state transition probabilites.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HMMModel$setStateTransitions(state.transitions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state.transitions}}{A matrix of state transition probabilites.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{M <- HMMModel$new("Ex", c("state1", "state2"), c("signalA", "signalB", "signalC"))
M$setStateTransitions(matrix(c(0.4, 0.6, 0.6, 0.4), byrow = TRUE, ncol = 2))
M$state.transitions
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setSignalProbability"></a>}}
\if{latex}{\out{\hypertarget{method-setSignalProbability}{}}}
\subsection{Method \code{setSignalProbability()}}{
Set signal probabilites.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HMMModel$setSignalProbability(signal.probabilities)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signal.probabilities}}{A matrix of signal probabilites per state.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{M <- HMMModel$new("Ex", c("state1", "state2"), c("signalA", "signalB", "signalC"))
M$setSignalProbability(matrix(c(0.4, 0.4, 0.2, 0.1, 0.1, 0.8), byrow = TRUE, ncol = 3))
M$signal.probabilities
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print information for the HMM.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HMMModel$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{-}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HMMModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
