% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_decoding.R
\name{l_decoding}
\alias{l_decoding}
\title{Local state decoding}
\usage{
l_decoding(obs, delta, Gamma, p)
}
\arguments{
\item{obs}{Observations.}

\item{delta}{m-dimensional vector. Initial distribution of Markov chain}

\item{Gamma}{m x m matrix. Transition probs for Markov chain}

\item{p}{Function returning the evaluated density given the state.}
}
\value{
List of most likely states and decoding probabilities.
}
\description{
A function for local decoding of HMM's. Not expected to be called by user. Called by the general HMM function.
}
\keyword{internal}
