% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LearnBaumWelch_B}
\alias{LearnBaumWelch_B}
\title{Model parameter learning - Baum-Welch Algorithm - B}
\usage{
LearnBaumWelch_B(A, B, PI, O, e = 0.05)
}
\arguments{
\item{A}{initial transition probability}

\item{B}{initial observation probability}

\item{PI}{initial probability distribution}

\item{O}{observation sequence}

\item{e}{threshold value}
}
\value{
B parameter estimation
}
\description{
Model parameter learning - Baum-Welch Algorithm - B
}
