% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_LH.R
\name{LH}
\alias{LH}
\title{Likelihood of the Hidden Markov Model}
\usage{
LH(factor, x, m, L1, L2, L3 = NULL, L4 = NULL, L5 = NULL)
}
\arguments{
\item{factor}{Input of variables that are unrestricted}

\item{x}{a sample of a Mixed Model}

\item{m}{the number of states}

\item{L1}{likelihood of the first hidden state}

\item{L2}{likelihood of the second hidden state}

\item{L3}{optional. likelihood of the third hidden state}

\item{L4}{optional. likelihood of the 4th hidden state}

\item{L5}{optional. likelihood of the 5th hidden state}
}
\value{
negative Likelihood
}
\description{
This function calculates the log-likelihood of the HMM model by normalizing the forward-probabilities.
}
\details{
This function computes the log-likelihood of the forward probabilities of the HMM given the factor values
Due to the fact that the Sigma and Gamma matrices are in a form restricted, the transformation function is used to 
transform the factor variables in suitable canidates. 

The likelihood is constructed with normalizing the alpha vectors.
}
