% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{load_jams}
\alias{load_jams}
\title{Load jams from disk}
\usage{
load_jams(jams_path)
}
\arguments{
\item{jams_path}{Path to the location of the unzipped thisismyjam folder}
}
\value{
returns a matrix of jams containing the following columns
\item{jam_id}{Identification number of a jam}
\item{user_id}{Identification number of the user that created the jam}
\item{artist}{Artist of the jam}
\item{title}{Title of song of the jam}
\item{creation_date}{Title of song of the jam}
}
\description{
This function loads the jams.tsv file from the THIS IS MY JAM dataset into R.
Furthermore users who have at least one corrupted value will be deleted.
}
\examples{
jams <- load_jams("S:/")
}
