% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_HMM_EM.R
\name{multi_HMM_EM}
\alias{multi_HMM_EM}
\title{Fitting a Hidden Markov Model via Expectation-Maximisation Algorithm}
\usage{
multi_HMM_EM(x, theta, m, L1, L2, L3 = NULL, L4 = NULL, L5 = NULL,
  iterations = NULL, DELTA = NULL)
}
\arguments{
\item{x}{a sample of a Hidden Markov Model}

\item{theta}{initial parameters for the estimation of the likelihood 
parameters. See details for more information}

\item{m}{the number of states}

\item{L1}{likelihood of the first hidden state}

\item{L2}{likelihood of the second hidden state}

\item{L3}{optional. likelihood of the third hidden state}

\item{L4}{optional. likelihood of the 4th hidden state}

\item{L5}{optional. likelihood of the 5th hidden state}

\item{iterations}{optional. number of iterations for the EM-Algorithm}

\item{DELTA}{optional. stop criterion for the EM-Algorithm}
}
\value{
The estimated parameters are rounded by 3 decimals and returned in a 
list.
}
\description{
Estimation of the transition probabilites, the initial state
probabilites and the hidden state parameters of a Hidden Markov Model
by using the Baum-Welch Algorithm.
}
\details{
This function estimates the hidden states of the Hidden Markov Model
with the help of the Baum Welch algorithm. The function iteratively applies 
both estimation and maximisation steps to arrive at the predicted parameters.
When the maximal difference between present and prior parameter is abitrarily 
small (defined with DELTA) the iteration stops.

This function is able to calculate with multiple Theta values for the 
individual likelihoods. For each likelihood the right number initial starting
parameter has to be set in order to compute the estimation of the 
corresponding Thetas.
For each Likelihood the starting values must be in the format of a vector, 
which is then saved as a list element. 

e.g.: theta[[i]] <- c(parameter1, parameter2, ...)

The function then extracts the right number of parameters per likelihood and 
optimizes the values.
}
