% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_LH.R
\name{multi_LH}
\alias{multi_LH}
\title{Likelihood of the Hidden Markov Model}
\usage{
multi_LH(factor, x, m, L1, L2, L3 = NULL, L4 = NULL, L5 = NULL,
  start_index)
}
\arguments{
\item{factor}{Input of variables that are unrestricted}

\item{x}{a sample of a Hidden Markov Model}

\item{m}{the number of states}

\item{L1}{likelihood of the first hidden state}

\item{L2}{likelihood of the second hidden state}

\item{L3}{optional. likelihood of the third hidden state}

\item{L4}{optional. likelihood of the 4th hidden state}

\item{L5}{optional. likelihood of the 5th hidden state}

\item{start_index}{index parameter to assign the Thetas to their corresponding 
individual hidden state likelihood.}
}
\value{
negative Likelihood
}
\description{
This function calculates the log-likelihood of the HMM mode.
For this, the scaled forward probabilities are computed. In respect of the LH
function, the multi_LH function incorporates the multi Thetas for the 
indivudal likelihoods.
}
\details{
This function computes the log-likelihood of the forward 
probabilities of the HMM. Given the fact that the inputed factor vector is 
not restricted, we need to apply a transformation to transform the factor
variables to our suitable canditates Delta, Gamma and Theta. For this we apply 
the function trans().

The likelihood is constructed using the scaling/normalizing of the forward 
probabilities such that for each time point the sum of the foward probability
is equal to one. The scaling is neccesary to tackle the underflow problem, 
that arrises with an increasing sample size.

The multi_LH function only differes in the computation of the likelihood 
probability vector p, because we imput multiple Thetas in the corresponding 
individual likelihoods.
}
