% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_theta_estimation.R
\name{multi_theta_estimation}
\alias{multi_theta_estimation}
\title{Multi Theta Estimation function}
\usage{
multi_theta_estimation(m, N, u, x, theta, L1, L2, L3 = NULL, L4 = NULL,
  L5 = NULL, start_index)
}
\arguments{
\item{m}{number of likelihoods}

\item{N}{length of the supplied dataset}

\item{u}{output matrix of the u_function}

\item{x}{a sample of a Hidden Markov Model}

\item{theta}{Theta vector}

\item{L1}{likelihood of the first hidden state}

\item{L2}{likelihood of the second hidden state}

\item{L3}{optional. likelihoods of the third hidden state}

\item{L4}{optional. likelihoods of the 4th hidden state}

\item{L5}{optional. likelihoods of the 5th hidden state}

\item{start_index}{index paramter to assign the right amount of thetas 
to the likelihoods}
}
\value{
returns the corresponding likelihood
}
\description{
This function calculates part of the global log-likelihood that 
is only dependent on the Theta value. Due to its proportionality, it is 
therefore optimal for the maximisation of the Theta values and will be used 
by the EM-algorithm. For the multi_HMM_EM(), small changes where made to 
calculate the index of the Theta alues right.
}
\details{
For more detailed explanation we recommend the source Hidden Markov 
Models for Times Series by Walter Zucchini, Iain MacDonald & Roland Langrock, 
especially page 72.
}
