% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_HMM_EM.R
\name{multiHMM2}
\alias{multiHMM2}
\title{Fitting a Hidden Markov Model via Expectation-Maximisation Algorithm}
\usage{
multiHMM2(x, theta, m, L1, L2, L3 = NULL, L4 = NULL, L5 = NULL,
  iterations = NULL, delta = NULL)
}
\arguments{
\item{x}{a sample of a Mixed Model}

\item{theta}{initial parameters for the estimation of the likelihood parameters. See details for more information}

\item{m}{the number of states}

\item{L1}{likelihood of the first hidden state}

\item{L2}{likelihood of the second hidden state}

\item{L3}{optional. likelihood of the third hidden state}

\item{L4}{optional. likelihood of the 4th hidden state}

\item{L5}{optional. likelihood of the 5th hidden state}

\item{iterations}{optional. number of iterations for the EM-Algorithm}

\item{delta}{optional. stop criterion for the EM-Algorithm}
}
\value{
The estimated parameters are rounded by 3 decimals and returned in a list.
}
\description{
Estimation of the transition probabilites, the initial state probabilites and the hidden state parameters of a Hidden Markov Model
by using the Baum-Welch Algorithm.
}
\details{
This function estimates the hidden states of the Hidden Markov Model with the
help of the Baum Welch algorithm. The function iteratively applies both estimation and
maximisation steps to arrive at the predicted parameters. When the maximum difference between
the functions is abitrarily small (see delta) the iteration stops.

For each individual likelihood an initial starting parameter has to be set in order to compute the estimation of the corresponding
Thetas. Each groupe of parameters is placed in a seperate element of the theta list as a vector e.g.:
theta[[i]] <- c(parameter1,parameter2,...)
}
