% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_HMM_DM.R
\name{multiHMM3}
\alias{multiHMM3}
\title{Fitting a Hidden Markov Model via the Direct Maximisation}
\usage{
multiHMM3(x, theta, m, L1, L2, L3, L4, L5)
}
\arguments{
\item{x}{a sample of a Hidden Markov Model}

\item{theta}{initial parameters for the estimation of the likelihood 
parameters. See details for more information.}

\item{m}{the number of states}

\item{L1}{likelihood of the first hidden state}

\item{L2}{likelihood of the second hidden state}

\item{L3}{optional. likelihood of the third hidden state}

\item{L4}{optional. likelihood of the 4th hidden state}

\item{L5}{optional. likelihood of the 5th hidden state}
}
\value{
The estimated parameters are rounded by 3 decimals and returned in a 
list.
}
\description{
Estimation of the transition probabilites,  
the initial state probabilites and the hidden state parameters of a Hidden 
Markov Model by using the Direct Maximisation of the global log-likelihood.
}
\details{
This function estimates the Hidden Markov states by maximising the 
normalized log-likelihood of the forward propabilities. Due to the fact that
both the Gamma matrix as well as the Delta vector have some constraints, the 
function first applies some restrictions and then uses the base-R maximisation
to gain the most likely variables. 

This function is able to calculate with multiple Theta values for the 
individual likelihoods. For each likelihood the right number initial starting
parameter has to be set in order to compute the estimation of the 
corresponding Thetas.
For each Likelihood the starting values must be in the format of a vector, 
which is then saved as a list element. 

e.g.: theta[[i]] <- c(parameter1, parameter2, ...)

The function then extracts the right number of parameters per likelihood and 
optimizes the values.
}
