% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.hmm}
\alias{plot.hmm}
\title{Plot method for class 'hmm'.}
\usage{
\method{plot}{hmm}(
  model,
  legend_position = "topright",
  xlab = "Time",
  ylab = "Observed emissions",
  ...
)
}
\arguments{
\item{model}{Class \code{hmm} model.}

\item{legend_position}{Position of legend in plots.}

\item{xlab}{Text for x-axis. Default 'Time'.}

\item{ylab}{Text for y-axis. Default 'Observed emissions'.}

\item{...}{}
}
\value{
Yields four plots stored in a list:\tabular{ll}{
    \tab  \cr
   \code{p1} \tab Plot of observed emissions by most likely hidden state (Global) \cr
   \code{p2} \tab Plot of observed emissions by most likely hidden state (Local) \cr
   \code{p3} \tab Plot of probs. used in Viterbi algorithm (Global) \cr
   \code{p4} \tab Plot of state probabilities (Local) \cr
}
}
\description{
Plot method for class 'hmm'.
}
\examples{
#Annual counts of earthquakes magnitude 7 or greater, 1900-2006.
#Source:
#Earthquake Data Base System of the U.S. Geological Survey, National
#Earthquake Information Center, Golden CO

obs <- read.table("http://www.hmms-for-time-series.de/second/data/earthquakes.txt")[,2]

#Define 2-state HMM with distribution family Poisson.
Gamma <- matrix(rep(.5, 2**2), ncol = 2); delta <- rep(.5, 2); lambda <- c(10, 20)
my_hmm_model <- hmm::hmm(obs, Gamma, delta, dist = "poisson", lambda = lambda)

#Decoding plots.
plot(my_hmm_model)

}
