% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_methods.R
\name{print.hmm}
\alias{print.hmm}
\title{Printing hidden Markov models}
\usage{
\method{print}{hmm}(x, ...)
}
\arguments{
\item{x}{Object of class \code{hmm}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for class \code{hmm}.
Prints a short summary of the involved parameters, namely the initial distribution, transition matrix and emission parameters.
}
\keyword{internal}
