% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_methods.R
\name{print.summary.hmm}
\alias{print.summary.hmm}
\title{Summarizing Hidden Markov Models}
\usage{
\method{print}{summary.hmm}(x, ...)
}
\arguments{
\item{...}{Additional arguments passed to \link[hmm]{fitted.hmm}}
}
\value{
If no data is available, essentially returns just the object ordered in a more reasonable manner and without internal functions.
Otherwise, also returns fitted values and residuals.
}
\description{
Summary method for class "\code{hmm}"
}
\details{
The function prints a summary of the parameters of the given model in a more human-readable format.
If fitted, it also provides model fitness statistics in the form of log-likelihood, AIC and BIC.
}
\examples{
TBD
}
