% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_methods.R
\name{residuals.hmm}
\alias{residuals.hmm}
\title{Residuals of Hidden Markov Models}
\usage{
\method{residuals}{hmm}(object, ...)
}
\arguments{
\item{object}{Object of type \code{hmm}.}

\item{...}{Other arguments passed to \link[hmm]{fitted.hmm}.}
}
\value{
Vector of residuals of the observed emissions.
}
\description{
Returns residuals of the observed emissions.
}
\details{
If the distribution family is one of the included standard families (Poisson, normal, binomial or exponential), the expected value will be the plug-in estimate using the fitted parameters of each state.
Otherwise, the empirical mean will be used as the expected value.
}
\examples{
## Continuation of Earthquake example
\dontshow{example(hmm)}
residuals(hmm.EQ)

## Continuation of simulated mixture data
\dontshow{example(simulate.hmm)}
hmm.mixture_X <- hmm(X.mixture, Gamma=Gamma, delta=delta, lls=lls, param_lls=param, lls_mle=lls_mle, rdist=rdist, estimate=FALSE)
residuals(hmm.mixture_X)
}
