% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{save_model}
\alias{save_model}
\title{Saves Hidden Markov model to disk}
\usage{
save_model(model, model_name)
}
\arguments{
\item{model}{a hidden markov model for music information retrieval created by MIR_hmm class}

\item{model_name}{name of the model on disk}
}
\description{
This function saves a hidden markov model as serialised object to disk
}
\examples{
hmm <- load_model("2_state_hmm")

save_model(hmm, "2_state_hmm")
}
