% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_hmm.R
\name{sim_hmm}
\alias{sim_hmm}
\title{Simulating Hidden Markov Models}
\usage{
sim_hmm(n, delta, Gamma, rdists, include_state = FALSE)
}
\arguments{
\item{n}{Integer, number of observations to simulate.}

\item{delta}{m-dimensional vector. Initial distribution of Markov chain}

\item{Gamma}{m x m matrix. Transition probs for Markov chain}

\item{rdists}{List of functions f_i, where f_i(k) generates k i.i.d. copies of X|C=i}

\item{include_state}{Logical. Denotes whether or not the hidden states should be included. If TRUE, they are appended to the end of the HMM.}
}
\value{
Vector of simulated HMM
}
\description{
A function for simulating HMM's. Not expected to be called by user, use instead the 'simulate' method for hmm objects.
}
