% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_methods.R
\name{summary.hmm}
\alias{summary.hmm}
\title{Summarizing Hidden Markov Models}
\usage{
\method{summary}{hmm}(object, ...)
}
\arguments{
\item{object}{An object of class "\code{hmm}", usually a result of call to \link[hmm]{hmm}.}

\item{...}{Additional arguments passed to \link[hmm]{fitted.hmm}}
}
\value{
If no data is available, essentially returns just the object ordered in a more reasonable manner and without internal functions.
Otherwise, also returns fitted values and residuals.
}
\description{
Summary method for class "\code{hmm}"
}
\details{
The function prints a summary of the parameters of the given model in a more human-readable format.
If fitted, it also provides model fitness statistics in the form of log-likelihood, AIC and BIC.
}
\examples{
## Continuation of Earthquake example
\dontshow{example(hmm)}
summary(hmm.EQ)

## Continuation of simulated mixture data
\dontshow{example(simulate.hmm)}
hmm.mixture_X <- hmm(X.mixture, Gamma=Gamma, delta=delta, lls=lls, param_lls=param, lls_mle=lls_mle, rdist=rdist, estimate=FALSE)
summary(hmm.mixture_X)
}
