% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_estimation.R
\name{theta_estimation}
\alias{theta_estimation}
\title{Theta Estimation function}
\usage{
theta_estimation(m, N, u, x, theta, L1, L2, L3 = NULL, L4 = NULL,
  L5 = NULL)
}
\arguments{
\item{m}{number of likelihoods}

\item{N}{length of the supplied dataset}

\item{u}{output matrix of the u_function}

\item{x}{a sample of a Mixed Model}

\item{theta}{theta vector}

\item{L1}{likelihood of the first hidden state}

\item{L2}{likelihood of the second hidden state}

\item{L3}{optional. likelihoods of the third hidden state}

\item{L4}{optional. likelihoods of the 4th hidden state}

\item{L5}{optional. likelihoods of the 5th hidden state}
}
\value{
returns the corresponding likelihood
}
\description{
This function calculates the third part of the overall likelihood function. 
This likelihood part focusses on the hidden states parameters and will be optimized later on.
It is estimated as a sum over all time periods and states of the vector u times the corresponding log-likelihood.
}
