% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hilfvektor v.R
\name{v_function}
\alias{v_function}
\title{v-function}
\usage{
v_function(m, N, alpha, beta, p, L, Gamma, set)
}
\arguments{
\item{m}{number of Likelihoods}

\item{N}{length of the supplied dataset}

\item{alpha}{alpha matrix}

\item{beta}{beta matrix}

\item{p}{vector of Likelihood probabilities of dataset}

\item{L}{Likelihood of alpha*beta'}

\item{Gamma}{Gamma matrix}

\item{set}{index vector to align the p vector}
}
\value{
returns the coresponding values BUT due to further calculation steps
the m x m matrix of v-values for each datapoint t is 
is returned as one row in the output matrix.
}
\description{
This function calculates the v-function which is part of the estimation step. 
Each element in v aligns with the conditional expectation to 
reach state k from the previous state j given your datapoint x(t)
}
