% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi.R
\name{viterbi}
\alias{viterbi}
\title{viterbi}
\usage{
viterbi(obs, delta, Gamma, p)
}
\arguments{
\item{obs}{Observations.}

\item{delta}{m-dimensional vector. Initial distribution of Markov chain}

\item{Gamma}{m x m matrix. Transition probs for Markov chain}

\item{p}{Function returning the evaluated density given the state.}
}
\value{
List of most likely states and decoding probabilities.
}
\description{
A function for global decoding of HMM's. Not expected to be called by user. Called by the general HMM function.
}
