% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Viterbi_cpp}
\alias{Viterbi_cpp}
\title{Prediction problem - Viterbi Algorithm}
\usage{
Viterbi_cpp(A, B, PI, V, Q, obs)
}
\arguments{
\item{A}{initial transition probability}

\item{B}{initial observation probability}

\item{PI}{initial probability distribution}

\item{V}{observation set}

\item{Q}{state set}

\item{O}{observation sequence}
}
\value{
Optimal path
}
\description{
Prediction problem - Viterbi Algorithm
}
