% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{add_genres_to_jams}
\alias{add_genres_to_jams}
\title{Add genre to jams}
\usage{
add_genres_to_jams(jams_clean, LFM_path)
}
\arguments{
\item{jams_clean}{return value of the loadJams() function}

\item{LFM_path}{Path to the location of the unzipped LFM folder}
}
\value{
table of jams with attached genre
}
\description{
This function loads the genres of the artist from the XXX dataset and merge them with the jams recieved from loadJams() function
}
