% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta_function}
\alias{beta_function}
\title{Normalized Backward Probability function}
\usage{
beta_function(m, N, Gamma, p, set, c)
}
\arguments{
\item{m}{number of likelihoods}

\item{N}{length of the supplied dataset}

\item{Gamma}{Gamma matrix}

\item{p}{vector of likelihood probabilities of the dataset}

\item{set}{index vector to align the p vector}

\item{c}{weights provided by the alpha calculation}
}
\value{
Returns the beta as a matrix with the dimensions m as rows
and the datapoints t as columns.
}
\description{
This function calculates the normalized backward probability of the HMM states. It is important
to recall, that the provided weights come from the calculation of the forward probabilities one step ahead
In the general literatur also revered to as beta matrix.
}
