% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_functions.R
\name{beta_function}
\alias{beta_function}
\title{Normalized Backward Probability function EM}
\usage{
beta_function(m, N, Gamma, p, set, weight)
}
\arguments{
\item{m}{number of likelihoods}

\item{N}{length of the supplied dataset}

\item{Gamma}{Gamma matrix}

\item{p}{vector of likelihood probabilities of the dataset}

\item{set}{index vector to align the p vector}

\item{weight}{weights provided by the alpha calculation}
}
\value{
Returns the beta as a matrix with the dimensions m as rows
and the datapoints t as columns.
}
\description{
This function calculates the normalized backward probability 
of the HMM states. For this the function needs the weights of the forward
probability, which have to be inputed as parameter weight. 
In the general literatur also revered to as beta matrix.
}
