% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Decoding.R
\name{decode}
\alias{decode}
\title{Local and global decoding}
\usage{
decode(x, m, L1, L2, L3 = NULL, L4 = NULL, L5 = NULL, gamma, delta,
  theta, multi = FALSE)
}
\arguments{
\item{x}{a sample of a Mixed Model}

\item{m}{the number of states}

\item{L1}{likelihood of the first hidden state}

\item{L2}{likelihood of the second hidden state}

\item{L3}{optional. likelihood of the third hidden state}

\item{L4}{optional. likelihood of the 4th hidden state}

\item{L5}{optional. likelihood of the 5th hidden state}

\item{gamma}{estimated gamma matrix from previous estimation}

\item{delta}{estimated delta vector from previous estimation}

\item{theta}{estimated theta values from previous estimation}

\item{multi}{parameter that checks, wether the function is used for multiple thetas or not}
}
\description{
Calculation of local and global decoding given the dataset and the 
coresponding parameters of the HMM
}
