% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{em}
\alias{em}
\title{em}
\usage{
em(
  obs,
  gamma,
  delta,
  lls,
  param_lls,
  lls_mle,
  epsilon = 1e-05,
  max_iter = 1000,
  ...
)
}
\arguments{
\item{obs}{The observed data}

\item{gamma}{Initial value of transition matrix}

\item{delta}{Initial value of initial distribution}

\item{lls}{List of marginal densities}

\item{param_lls}{List of parameters for marginal densities}

\item{lls_mle}{List of functions that generate MLE's}

\item{epsilon}{Threshold to stop algorithm when difference in log-likelihood is less than this}

\item{max_iter}{Maximum number of iterations before stopping}

\item{...}{Additional arguments, mainly here to avoid "unused argument"-errors.}
}
\value{
A list of log-likelihoods, parameter estimates and number of iterations.
}
\description{
Calculates approximate MLE's of all involved parameters, i.e. initial distribution probabilities, transition probabilities and whichever parameters are required for the marginal distribution of the emissions.
Not expected to be called by the user directly, instead being called during the initialization of an \code{hmm} object.
}
