% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_methods.R
\name{fitted.hmm}
\alias{fitted.hmm}
\title{Fitted values for Hidden Markov Models}
\usage{
\method{fitted}{hmm}(x, method = "local", ...)
}
\arguments{
\item{x}{Object of type \code{hmm}.}

\item{method}{Method for state prediction. One of either 'local' (default, resulting in local decoding), 'viterbi' or 'global' (both resulting in global decoding using Viterbi's algorithm).}

\item{...}{Other arguments.}
}
\value{
List of predicted states and emission values.
}
\description{
Outputs predicted values of Hidden Markov Model in the form of a list where the first element is a vector of state predictions and second element is a vector of emission predictions (expected value corresponding predicted state).
}
\details{
If the distribution family is one of the included standard families (Poisson, normal, binomial or exponential), the expected value will be the plug-in estimate using the fitted parameters of each state.
Otherwise, the empirical mean will be used as the expected value.
}
\examples{
# Continuation of Earthquake data example
quakes <- read.table("http://www.hmms-for-time-series.de/second/data/earthquakes.txt")$V2
Gamma <- rbind(c(0.9, 0.1), c(0.1, 0.9))
delta <- c(1, 1)/2
lambda <- c(10, 30)
M <- hmm(quakes, Gamma, delta, dist='poisson', lambda=lambda)

fitted.values(M)
}
