% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_logprobabilities.R,
%   R/backward_probabilities.R, R/forward_logprobabilities.R,
%   R/forward_probabilities.R
\name{backward_logprobabilities}
\alias{backward_logprobabilities}
\alias{backward_probabilities}
\alias{forward_logprobabilities}
\alias{forward_probabilities}
\title{Forward and backward (log-)probabilities}
\usage{
backward_logprobabilities(x, Gamma, p)

backward_probabilities(x, Gamma, p)

forward_logprobabilities(x, Gamma, p, delta)

forward_probabilities(x, Gamma, p, delta)
}
\arguments{
\item{x}{Vector of observed emissions}

\item{Gamma}{Matrix (m x m) of transition probabilities of the underlying Markov chain}

\item{p}{A function where p(i, x)=P(X_t=x | Y_t=i)}
}
\description{
Calculate matrices of forward and backward (log-)probabilities for a given data set and parameters in a hidden Markov model.
Essentially just a wrapper for the true function written in C++.
Useful for avoiding underflow (by using the log-exp-sum-trick) when the HMM is large.
Not expected to be called by the user, instead only being used during the EM-algorithm.
}
\details{
Let (X, Y)_i, i=1,...,n constitute a HMM, that is, Y_1,...,Y_n is a Markov chain with states 1,...,m and the X_i's are dependent only through the Y_i's with common conditional distributions X_i | Y_i=j = P_j for all i=1,...,n. This function then generates a matrix where the i,j'th entry is log P(X_j+1=x_j+1, ..., X_n=x_n | Y_j=i), where x_1,...,x_n are the observed values (emissions) of X.
}
\section{Functions}{
\itemize{
\item \code{backward_logprobabilities}: Matrix of backwards log-probabilities

\item \code{backward_probabilities}: Matrix of backward probabilities

\item \code{forward_logprobabilities}: Matrix of forward log-probabilities

\item \code{forward_probabilities}: Matrix of forward probabilities
}}

\keyword{internal}
