\encoding{latin1}
\name{HMM}
\alias{HMM}
\docType{package}
\title{
HMM - Hidden Markov Models
}
\description{
Modelling, analysis and inference with discrete time and discrete space Hidden Markov Models.
}
\details{
\tabular{ll}{
Package: \tab HMM - Rpackage\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-01-10\cr
License: \tab GPL version 2 or later\cr
Maintainer: \tab Scientific Software Development - Dr. Lin Himmelmann, www.linhi.com\cr
}
}
\keyword{package}