% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{hmm}
\alias{hmm}
\title{Fitting Hidden Markov Models}
\usage{
hmm(x, Gamma, delta, dist = NULL, ..., estimate = !is.null(x))
}
\arguments{
\item{x}{Numeric vector of observed emissions. Can be NULL if no estimation is desired.}

\item{Gamma}{Initial value of transition probability matrix for underlying Markov chain.}

\item{delta}{Numeric vector of probabilities of the initial distribution of the Markov chain.}

\item{dist}{Distribution family of emissions. Can be one of 'poisson', 'normal', 'binom', 'exponential' or NULL. If NULL, user must provide own functions for densities, MLEs and random generation (see details)}

\item{...}{Parameters of emission distribution as well as parameters for EM-algorithm. See details.}

\item{estimate}{Logical variable whether or not parameters of model (Gamma, delta and emission parameters) should be estimated by EM-algorithm. Defaults to TRUE when data is available.}
}
\value{
An object of class 'hmm'.
The 'hmm'-class is equipped with the following methods: \code{summary}, \code{plot}, \code{fitted.values}, \code{residuals}, \code{logLik} and \code{simulate}.
An object of class 'hmm' is a list containing at least the following components:\tabular{ll}{
    \tab  \cr
   \code{m} \tab Number of hidden states. \cr
   \code{dist} \tab Distribution family of emissions. \cr
   \code{Gamma} \tab The transition probability matrix of the underlying Markov chain. \cr
   \code{delta} \tab The initial distribution of the underlying Markov chain. \cr
   \code{parameters} \tab List of parameters for the emission distribution family. \cr
   \code{rdists} \tab List of functions used for simulating. Mostly for internal use. \cr
}


If \code{x} is not \code{NULL}, it will also include:\tabular{ll}{
    \tab  \cr
   \code{x} \tab Vector of observations. \cr
   \code{n} \tab Number of observations. \cr
   \code{logLik} \tab Log-likelihood of parameters given the observed data. \cr
   \code{AIC} \tab AIC of the model. \cr
   \code{BIC} \tab BIC of the model. \cr
}


Finally, if estimation is performed, it will also include the following:\tabular{ll}{
    \tab  \cr
   \code{EMlogLik} \tab Vector of log-likelihoods at each iteration in the EM-algorithm. \cr
   \code{n_iter} \tab Number of iterations performed in the EM-algorithm. \cr
}
}
\description{
\code{hmm} is used to fit (or simply define) Hidden Markov Models.
Most commonly, it will be used to simply fit a vector of observed emissions which come from a (known) common distribution family to a hidden Markov model, where the parameters of the underlying Markov chain as well as the emission distribution family is then fitted using the EM-algorithm.

Alternatively, one can use custom distribution families by providing various functions (see details).

Finally, one can also disable optimization altogether to simply obtain an 'hmm'-object which can then be assessed through various standard methods (see below).
}
\details{
TODO
}
\examples{
TODO
}
\seealso{
TODO
}
