% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{HMM}
\alias{HMM}
\title{Fitting a Hidden Markov Model}
\usage{
HMM(x, m, methode = "DM", L1, L2, L3 = NULL, L4 = NULL, L5 = NULL,
  iterations = NULL, delta = NULL)
}
\arguments{
\item{x, }{a sample of a Mixed Model}

\item{m, }{the number of states}

\item{methode, }{choose between two different methodes: "DM" as default, alternative "EM"}

\item{L1, }{likelihood of the first hidden state}

\item{L2, }{likelihood of the second hidden state}

\item{iterations, }{optional. Number of iterations for the EM-Algorithm}

\item{delta, }{optional. Stop criterion for the EM-Algorithm}

\item{L3-L5, }{optional. likelihoods of the third, 4th and 5th hidden state}
}
\value{
The estimated parameters are rounded by 3 decimals and returned in a list.
}
\description{
Estimation of the transition probabilites, the initial state probabilites and the hidden state parameters of a Hidden Markov Model
by using an EM-Algorithm
}
\details{
This package is designed to estimate the hidden states of a HMM-Model, given the underlyning Likelihoods
of each state. It is important to support at least two Likelyhoods (L1,L2) for the function, which both
depend on a unknown parameter theta. 

In the methode parameter the underlying estimation function is selected. With DM the HMM-function will
estimate the parameters via a direct maximisation of the given Likelihoods.
If EM is selected the HMM-function will use a Baum-Welch estimation algorithm to to compute the different states and the
estimation of the underlying parameters. 

For more detailed explenation we recomend the source Hidden Markov Models for Times Series 
by Walter Zucchini, Iain MacDonald & Roland Langrock


The underlying functions are the HMM2 for the EM-algorithm and the HMM3 for the direct Maximisation
}
