% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM estimation with EM.R
\name{HMM2}
\alias{HMM2}
\title{Fitting a Hidden Markov Model via Expectation-Maximisation algorithm}
\usage{
HMM2(x, m, L1, L2, L3 = NULL, L4 = NULL, L5 = NULL, iterations = NULL,
  delta = NULL)
}
\arguments{
\item{x, }{a sample of a Mixed Model}

\item{m, }{the number of states}

\item{L1, }{likelihood of the first hidden state}

\item{L2, }{likelihood of the second hidden state}

\item{iterations, }{optional. Number of iterations for the EM-Algorithm}

\item{delta, }{optional. Stop criterion for the EM-Algorithm}

\item{L3-L5, }{optional. likelihoods of the third, 4th and 5th hidden state}
}
\value{
The estimated parameters are rounded by 3 decimals and returned in a list.
}
\description{
Estimation of the transition probabilites, the initial state probabilites and the hidden state parameters of a Hidden Markov Model
by using the Baum-Welch Algorithm.
}
\details{
This function estimates the hidden states of the Hidden Markov Model with the
help of the Baum Welch algorithm. The function iterativly applies both estimation and
maximisation steps to arrive at the predicted parameters. When the maximum difference between
the functions is abitrarily small (see delta) the iteration stops.
}
