% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM estimation with direct maximisation.R
\name{HMM3}
\alias{HMM3}
\title{Fitting a Hidden Markov Model via the direct maximisation}
\usage{
HMM3(x, m, L1, L2, L3, L4, L5)
}
\arguments{
\item{x, }{a sample of a Mixed Model}

\item{m, }{the number of states}

\item{L1, }{likelihood of the first hidden state}

\item{L2, }{likelihood of the second hidden state}

\item{L3-L5, }{optional. likelihoods of the third, 4th and 5th hidden state}
}
\value{
The estimated parameters are rounded by 3 decimals and returned in a list.
}
\description{
Estimation of the transition probabilites, the initial state probabilites and the hidden state parameters of a Hidden Markov Model
by using the direct Maximisation of the Likelihoods
}
\details{
This function estimates the Hidden Markov states by maximising the normalized Log-Likelihood
of the forward propabilities. Due to the fact that bot the Gamma-matrix as well as the Sigma-matrix 
have some constrains, the function includes the restrictions within the function.
}
