% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{load_jams}
\alias{load_jams}
\title{Load jams from disk}
\usage{
load_jams(jams_path)
}
\arguments{
\item{jams_path}{Path to the location of the unzipped thisismyjam folder}
}
\value{
table with the jams
}
\description{
This function loads the jams.tsv file from the THIS IS MY JAM dataset into R.
Furthermore users who have at least one corrupted value will be deleted.
}
