% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{load_model}
\alias{load_model}
\title{Loads a Hidden Markov model from disk}
\usage{
load_model(model_name)
}
\arguments{
\item{model_name}{name of the model on disk}
}
\value{
a hidden markov model object
}
\description{
This function loads a hidden markov model from disk
}
\examples{
hmm <- load_model("2_state_hmm")
print(hmm)
}
