% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{MIR_hmm}
\alias{MIR_hmm}
\title{Class to initialise a hidden markov model for music information retrieval}
\usage{
MIR_hmm(training_data, K = 2, verbose = TRUE, tolerance = 0.001,
  itmax = 300)
}
\arguments{
\item{training_data}{a matrix of sequences}

\item{K}{number of hidden states of the hidden markov model for music information retrieval}

\item{verbose}{variable determinig if the EM algorithm steps should be printed}

\item{tolerance}{variable determing at which percentage change in log-likelihood the convergence citeria is met}

\item{itmax}{variable determinig how many steps the EM algorithm should run at maximum}
}
\value{
a hidden markov model for music information retrieval
}
\description{
This function initialise a hidden markov model for music information retrieval. This includes two hidden markov models.
One of the HMM package and one of the hmm.discnp package. hmm.discnp is used for fitting the data to the model while hmm
is used for forward function to predict next values
}
