% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_factor_functions.R
\name{multifactorHMM}
\alias{multifactorHMM}
\title{Fitting a Hidden Markov Model to multifactor Likelihoods}
\usage{
multifactorHMM(x, theta, m, method = "DM", L1, L2, L3 = NULL,
  L4 = NULL, L5 = NULL, iterations = NULL, delta = NULL,
  decoding = FALSE)
}
\arguments{
\item{x}{a sample of a Mixed Model}

\item{theta}{list with initial number of Likelihood parameters (see details)}

\item{m}{the number of states}

\item{method}{choose between two different methods: "DM" as default, alternative "EM"}

\item{L1}{likelihood of the first hidden state}

\item{L2}{likelihood of the second hidden state}

\item{L3}{optional. likelihood of the third hidden state}

\item{L4}{optional. likelihood of the 4th hidden state}

\item{L5}{optional. likelihood of the 5th hidden state}

\item{iterations}{optional. number of iterations for the EM-Algorithm}

\item{delta}{optional. stop criterion for the EM-Algorithm}

\item{decoding}{if parameter set TRUE the function returns the most probable paths via local and global decoding}
}
\value{
The estimated parameters are rounded by 3 decimals and returned in a list
}
\description{
Estimation of the transition probabilites, the initial state probabilites and the hidden state parameters of a Hidden Markov Model
by using the Direct Maximisation of the likelihood or the EM-Algorithm.
}
\details{
This package is designed to estimate the hidden states of a HMM-Model, given the underlying likelihoods
of each state. It is important to support at least two likelihoods (L1, L2) for the function, which both
depend on multiple unkown thetas.

For each individual likelihood a starting parameter has to be set in order to compute the estimation of the corresponding
Thetas. Each groupe of parameters is placed in a seperate element of the theta list as a vector e.g.:
theta[[i]] <- c(parameter1,parameter2,...)

In the method parameter the underlying estimation function is selected. With DM the HMM-function will
estimate the parameters via a direct maximisation of the given likelihoods.
If EM is selected the HMM-function will use a Baum-Welch estimation algorithm to compute the different states and the
estimation of the underlying parameters.

For more detailed explanation we recommend the source Hidden Markov Models for Times Series
by Walter Zucchini, Iain MacDonald & Roland Langrock.


The underlying functions are the multiHMM2 for the EM-Algorithm and the multiHMM3 for the Direct Maximisation
}
\seealso{
[HMM()]
}
