% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_HMM_DM.R
\name{multiHMM3}
\alias{multiHMM3}
\title{Fitting a Hidden Markov Model via the Direct Maximisation}
\usage{
multiHMM3(x, theta, m, L1, L2, L3, L4, L5)
}
\arguments{
\item{x}{a sample of a Mixed Model}

\item{theta}{initial parameters for the estimation of the likelihood parameters. See details for more information.}

\item{m}{the number of states}

\item{L1}{likelihood of the first hidden state}

\item{L2}{likelihood of the second hidden state}

\item{L3}{optional. likelihood of the third hidden state}

\item{L4}{optional. likelihood of the 4th hidden state}

\item{L5}{optional. likelihood of the 5th hidden state}
}
\value{
The estimated parameters are rounded by 3 decimals and returned in a list.
}
\description{
Estimation of the transition probabilites, the initial state probabilites and the hidden state parameters of a Hidden Markov Model
by using the Direct Maximisation of the likelihoods.
}
\details{
This function estimates the Hidden Markov states by maximising the normalized log-likelihood
of the forward probabilities. Due to the fact that bot the Gamma-matrix as well as the Sigma-matrix
have some constraints, the function includes the restrictions within the function.

For each individual likelihood an initial starting parameter has to be set in order to compute the estimation of the corresponding
Thetas. Each groupe of parameters is placed in a seperate element of the theta list as a vector e.g.:
theta[[i]] <- c(parameter1,parameter2,...)
}
