% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.hmm}
\alias{plot.hmm}
\title{Plotting Hidden Markov Models.}
\usage{
\method{plot}{hmm}(
  model,
  legend_position = "topright",
  xlab = "Time",
  ylab = "Observed emissions",
  ...
)
}
\arguments{
\item{model}{Class \code{hmm} model.}

\item{legend_position}{Position of legend in plots.}

\item{xlab}{Text for x-axis. Default 'Time'.}

\item{ylab}{Text for y-axis. Default 'Observed emissions'.}

\item{...}{}
}
\value{
Yields four plots stored in a list:\tabular{ll}{
    \tab  \cr
   \code{p1} \tab Plot of observed emissions by most likely hidden state (Global) \cr
   \code{p2} \tab Plot of observed emissions by most likely hidden state (Local) \cr
   \code{p3} \tab Plot of probs. used in Viterbi algorithm (Global) \cr
   \code{p4} \tab Plot of state probabilities (Local) \cr
}
}
\description{
Plotting Hidden Markov Models.
}
\examples{
## Continuation of Earthquake example
\dontshow{example(hmm)}
plot(hmm.EQ)

}
