% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{predict_next}
\alias{predict_next}
\title{Predict next observation}
\usage{
predict_next(MIR_hmm, obs)
}
\arguments{
\item{MIR_hmm}{a object of the s3 class MIR_hmm}

\item{obs}{a list of genres}
}
\value{
predicted genre
}
\description{
This function predicts the next observation for a given sequence of music genres. This is achieved
by calculating probability by each genre to be the next observation. The genre with the highest
probability is considered to be the next favorite genre of a user
}
\examples{
# Load MIR HMM
hmm <- load_model(2_state_hmm)
# Create sequence of observations
seq <- c("rock","heavy metal", "pop")
# Predict next genre
prediction <- predict_next(hmm,seq)
print(prediction)
}
