% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{split_data}
\alias{split_data}
\title{Split data into training and test dataset}
\usage{
split_data(data, split_ratio = 0.5)
}
\arguments{
\item{data}{a matrix containing one sequence per row}

\item{split_ratio}{a statement determining how many percent of the original dataset should be used as training dataset}
}
\value{
a list containing the training dataset at list index 1 and the test dataset at listindex 2
}
\description{
This function draws a random sample of a dataset as training dataset.
}
