% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_hmm.R
\name{summary.hmm}
\alias{summary.hmm}
\title{Summarizing Hidden Markov Models}
\usage{
\method{summary}{hmm}(object, ...)
}
\arguments{
\item{object}{An object of class "\code{hmm}", usually a result of call to}

\item{...}{Additional arguments affecting the summary produced}
}
\value{
TBD
}
\description{
Summary method for class "\code{hmm}"
}
\details{
The function prints a summary of the parameters of the given model in a more human-readable format.
If fitted, it also provides model fitness statistics in the form of log-likelihood, AIC and BIC.
}
\examples{
TBD
}
