% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test_model}
\alias{test_model}
\title{Predicts the last value of each sequence in the test dataset}
\usage{
test_model(trained_MIR_hmm, testdata)
}
\arguments{
\item{trained_MIR_hmm}{a trained model of the class MIR_hmm}

\item{testdata}{testdata as the result of the function split_data()}
}
\value{
a matrix of the predicted and the actual values
}
\description{
This function transforms a matrix that contains sequential data in TSE format
to STS format. In comparison with the TSE_TO_STS function of the TRAMINER package
this function keeps NA values and doesn't fills them up with the last non-NA value
}
