% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.R
\name{trans}
\alias{trans}
\title{Transformation function - DM}
\usage{
trans(factor, m)
}
\arguments{
\item{factor}{see details}

\item{m}{number of Likelihoods}
}
\value{
returns a matrix with the sigma,Gamma and theta matrix bound together (collumn wise)
}
\description{
Due to our single optimation problem we have to build constrains for Gamma and
Sigma to fullfill the requirements of probabilties.
}
\details{
In the direct maximisation the nlme()- minimisation function can not directly implement
the constrains of the parameter values. This function ensures that the the estimated parameters 
of the direct optimisation still fullfill their requirements. These are that the probabilities are
between zero and one and that the rows of gamma (as well as sigma) sum up to one.
For this transformation the probit model is used. 

Thus with the input factor containing the elements that determine Sigma and Gamma we need 
the following number of elements for each parameter:

Sigma vector (1 x m)  - (m-1)  elements required
Gamma matrix (m x m)  - m(m-1) elements required
Theta vector (1 x m)  - m      elements required

By this defintion the input factor vector has to contain the elements for Sigma, Gamma and Theta
in that order and has to have the dimension: (m+1)(m-1) + m
}
