% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{TSE_TO_STS2}
\alias{TSE_TO_STS2}
\title{Transform a TSE timesequence to a STS timesequence}
\usage{
TSE_TO_STS2(TSE_sequence)
}
\arguments{
\item{TSE_sequence}{a matrix in TSE format}
}
\value{
matrix in STS format
}
\description{
This function transforms a matrix that contains sequential data in TSE format
to STS format. This function works like the traminer one with the only difference that
empty cells won't be filled with the value of the last non NA cell but with NA values. Furthermore it doesn't have initial state
}
