% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{TSE_TO_STS2}
\alias{TSE_TO_STS2}
\title{Transform a TSE timesequence to a STS timesequence}
\usage{
TSE_TO_STS2(TSE_sequence)
}
\arguments{
\item{TSE_sequence}{A sequence in TSE format}
}
\value{
A matrix in STS format containing one sequence per row
}
\description{
This function transforms a matrix that contains sequential data in TSE (time stamped event sequences) format
to a matrix in STS (state sequences) format. This function works like the TSE_TO_STS() function of the TraMineRextras package but differs
in two points: If the passed dataset contains a sequence that is shorter than the longest sequence of the dataset TraMineRextras fills the
missing values with the last non missing value. This function will not do this and let the missing values be NA. Furthermore it doesn't have
a first state at the first value in the sequence. Please note that it is not necessary to use the function manually since the function is called
from the create_time_sequence() function.
}
