% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hilfvektor u.R
\name{u_function}
\alias{u_function}
\title{u-function}
\usage{
u_function(m, N, alpha, beta, L)
}
\arguments{
\item{m}{number of Likelihoods}

\item{N}{length of the supplied dataset}

\item{alpha}{alpha matrix}

\item{beta}{beta matrix}

\item{L}{Likelihood of alpha*beta'}
}
\value{
returns the coresponding u's
}
\description{
This function calculates the u-function which is part of the estimation step. 
Each element in u aligns with the conditional expectation to 
reach state j given your datapoint x(t)
}
